various cleanups
* refly::debug_print is now internal
* autodiff_aux.* integrated into autodiff.*
* continuation_dom/codom -> Lam/Pi::cn_dom/codom
* removed f_arg_ty macro
* improved docs
