+++ dialects/autodiff/passes/autodiff_eval.cpp
@@ -6,7 +6,6 @@
 
 #include "dialects/affine/affine.h"
 #include "dialects/autodiff/autodiff.h"
-#include "dialects/autodiff/auxiliary/autodiff_aux.h"
 #include "dialects/core/core.h"
 #include "dialects/mem/mem.h"
