+++ thorin/util/assert.h
@@ -4,7 +4,7 @@
 
 namespace thorin {
 
-/// @name Debug helpers
+/// @name Debug Helpers
 ///@{
 // see https://stackoverflow.com/a/65258501
 #ifdef __GNUC__ // GCC 4.8+, Clang, Intel and other compilers compatible with GCC (-std=c++0x or above)
